need to fix helpers in LCD
