+++ lcd_test/helpers_test.go
@@ -1,454 +1,43 @@
-package lcd_test
+package lcdtest
 
 import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
-	"os"
-	"path/filepath"
 	"regexp"
-	"sort"
 	"strings"
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/client"
+	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
+	"github.com/cosmos/cosmos-sdk/client/utils"
+	"github.com/cosmos/cosmos-sdk/crypto/keys"
+	"github.com/cosmos/cosmos-sdk/x/params"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
 
-	gapp "github.com/cosmos/gaia/app"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
-	"github.com/cosmos/cosmos-sdk/client/lcd"
 	"github.com/cosmos/cosmos-sdk/client/rpc"
-	"github.com/cosmos/cosmos-sdk/client/tx"
 	clienttx "github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/cosmos/cosmos-sdk/client/utils"
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/keys"
 	crkeys "github.com/cosmos/cosmos-sdk/crypto/keys"
-	"github.com/cosmos/cosmos-sdk/server"
-	"github.com/cosmos/cosmos-sdk/tests"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/rest"
 	"github.com/cosmos/cosmos-sdk/x/auth"
-	authrest "github.com/cosmos/cosmos-sdk/x/auth/client/rest"
 	txbuilder "github.com/cosmos/cosmos-sdk/x/auth/client/txbuilder"
-	"github.com/cosmos/cosmos-sdk/x/auth/genaccounts"
 	bankrest "github.com/cosmos/cosmos-sdk/x/bank/client/rest"
-	"github.com/cosmos/cosmos-sdk/x/crisis"
-	distr "github.com/cosmos/cosmos-sdk/x/distribution"
-	distrrest "github.com/cosmos/cosmos-sdk/x/distribution/client/rest"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
 	"github.com/cosmos/cosmos-sdk/x/gov"
 	govrest "github.com/cosmos/cosmos-sdk/x/gov/client/rest"
 	gcutils "github.com/cosmos/cosmos-sdk/x/gov/client/utils"
-	"github.com/cosmos/cosmos-sdk/x/mint"
-	mintrest "github.com/cosmos/cosmos-sdk/x/mint/client/rest"
-	paramsrest "github.com/cosmos/cosmos-sdk/x/params/client/rest"
 	paramscutils "github.com/cosmos/cosmos-sdk/x/params/client/utils"
 	"github.com/cosmos/cosmos-sdk/x/slashing"
 	slashingrest "github.com/cosmos/cosmos-sdk/x/slashing/client/rest"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 	stakingrest "github.com/cosmos/cosmos-sdk/x/staking/client/rest"
 
-	abci "github.com/tendermint/tendermint/abci/types"
-	tmcfg "github.com/tendermint/tendermint/config"
-	"github.com/tendermint/tendermint/crypto"
-	"github.com/tendermint/tendermint/crypto/ed25519"
-	"github.com/tendermint/tendermint/crypto/secp256k1"
-	"github.com/tendermint/tendermint/libs/cli"
-	dbm "github.com/tendermint/tendermint/libs/db"
-	"github.com/tendermint/tendermint/libs/log"
-	nm "github.com/tendermint/tendermint/node"
 	"github.com/tendermint/tendermint/p2p"
-	pvm "github.com/tendermint/tendermint/privval"
-	"github.com/tendermint/tendermint/proxy"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
-	tmrpc "github.com/tendermint/tendermint/rpc/lib/server"
-	tmtypes "github.com/tendermint/tendermint/types"
 )
-
-var cdc = codec.New()
-
-func init() {
-	codec.RegisterCrypto(cdc)
-}
-
-// makePathname creates a unique pathname for each test. It will panic if it
-// cannot get the current working directory.
-func makePathname() string {
-	p, err := os.Getwd()
-	if err != nil {
-		panic(err)
-	}
-
-	sep := string(filepath.Separator)
-	return strings.Replace(p, sep, "_", -1)
-}
-
-// GetConfig returns a Tendermint config for the test cases.
-func GetConfig() *tmcfg.Config {
-	pathname := makePathname()
-	config := tmcfg.ResetTestRoot(pathname)
-
-	tmAddr, _, err := server.FreeTCPAddr()
-	if err != nil {
-		panic(err)
-	}
-
-	rcpAddr, _, err := server.FreeTCPAddr()
-	if err != nil {
-		panic(err)
-	}
-
-	grpcAddr, _, err := server.FreeTCPAddr()
-	if err != nil {
-		panic(err)
-	}
-
-	config.P2P.ListenAddress = tmAddr
-	config.RPC.ListenAddress = rcpAddr
-	config.RPC.GRPCListenAddress = grpcAddr
-
-	return config
-}
-
-// CreateAddr adds an address to the key store and returns an address and seed.
-// It also requires that the key could be created.
-func CreateAddr(t *testing.T, name, password string, kb crkeys.Keybase) (sdk.AccAddress, string) {
-	var (
-		err  error
-		info crkeys.Info
-		seed string
-	)
-
-	info, seed, err = kb.CreateMnemonic(name, crkeys.English, password, crkeys.Secp256k1)
-	require.NoError(t, err)
-
-	return sdk.AccAddress(info.GetPubKey().Address()), seed
-}
-
-// CreateAddr adds multiple address to the key store and returns the addresses and associated seeds in lexographical order by address.
-// It also requires that the keys could be created.
-func CreateAddrs(t *testing.T, kb crkeys.Keybase, numAddrs int) (addrs []sdk.AccAddress, seeds, names, passwords []string) {
-	var (
-		err  error
-		info crkeys.Info
-		seed string
-	)
-
-	addrSeeds := AddrSeedSlice{}
-
-	for i := 0; i < numAddrs; i++ {
-		name := fmt.Sprintf("test%d", i)
-		password := "1234567890"
-		info, seed, err = kb.CreateMnemonic(name, crkeys.English, password, crkeys.Secp256k1)
-		require.NoError(t, err)
-		addrSeeds = append(addrSeeds, AddrSeed{Address: sdk.AccAddress(info.GetPubKey().Address()), Seed: seed, Name: name, Password: password})
-	}
-
-	sort.Sort(addrSeeds)
-
-	for i := range addrSeeds {
-		addrs = append(addrs, addrSeeds[i].Address)
-		seeds = append(seeds, addrSeeds[i].Seed)
-		names = append(names, addrSeeds[i].Name)
-		passwords = append(passwords, addrSeeds[i].Password)
-	}
-
-	return addrs, seeds, names, passwords
-}
-
-// AddrSeed combines an Address with the mnemonic of the private key to that address
-type AddrSeed struct {
-	Address  sdk.AccAddress
-	Seed     string
-	Name     string
-	Password string
-}
-
-// AddrSeedSlice implements `Interface` in sort package.
-type AddrSeedSlice []AddrSeed
-
-func (b AddrSeedSlice) Len() int {
-	return len(b)
-}
-
-// Less sorts lexicographically by Address
-func (b AddrSeedSlice) Less(i, j int) bool {
-	// bytes package already implements Comparable for []byte.
-	switch bytes.Compare(b[i].Address.Bytes(), b[j].Address.Bytes()) {
-	case -1:
-		return true
-	case 0, 1:
-		return false
-	default:
-		panic("not fail-able with `bytes.Comparable` bounded [-1, 1].")
-	}
-}
-
-func (b AddrSeedSlice) Swap(i, j int) {
-	b[j], b[i] = b[i], b[j]
-}
-
-// InitClientHome initialises client home dir.
-func InitClientHome(t *testing.T, dir string) string {
-	var err error
-	if dir == "" {
-		dir, err = ioutil.TempDir("", "lcd_test")
-		require.NoError(t, err)
-	}
-	// TODO: this should be set in NewRestServer
-	// and pass down the CLIContext to achieve
-	// parallelism.
-	viper.Set(cli.HomeFlag, dir)
-	return dir
-}
-
-// TODO: Make InitializeTestLCD safe to call in multiple tests at the same time
-// InitializeTestLCD starts Tendermint and the LCD in process, listening on
-// their respective sockets where nValidators is the total number of validators
-// and initAddrs are the accounts to initialize with some stake tokens. It
-// returns a cleanup function, a set of validator public keys, and a port.
-func InitializeTestLCD(t *testing.T, nValidators int, initAddrs []sdk.AccAddress, minting bool) (
-	cleanup func(), valConsPubKeys []crypto.PubKey, valOperAddrs []sdk.ValAddress, port string) {
-
-	if nValidators < 1 {
-		panic("InitializeTestLCD must use at least one validator")
-	}
-
-	config := GetConfig()
-	config.Consensus.TimeoutCommit = 100
-	config.Consensus.SkipTimeoutCommit = false
-	config.TxIndex.IndexAllTags = true
-
-	logger := log.NewTMLogger(log.NewSyncWriter(os.Stdout))
-	logger = log.NewFilter(logger, log.AllowError())
-
-	privVal := pvm.LoadOrGenFilePV(config.PrivValidatorKeyFile(),
-		config.PrivValidatorStateFile())
-	privVal.Reset()
-
-	db := dbm.NewMemDB()
-	app := gapp.NewGaiaApp(logger, db, nil, true, 0)
-	cdc = gapp.MakeCodec()
-
-	genesisFile := config.GenesisFile()
-	genDoc, err := tmtypes.GenesisDocFromFile(genesisFile)
-	require.Nil(t, err)
-	genDoc.Validators = nil
-	require.NoError(t, genDoc.SaveAs(genesisFile))
-
-	// append any additional (non-proposing) validators
-	var genTxs []auth.StdTx
-	var accs []genaccounts.GenesisAccount
-	for i := 0; i < nValidators; i++ {
-		operPrivKey := secp256k1.GenPrivKey()
-		operAddr := operPrivKey.PubKey().Address()
-		pubKey := privVal.GetPubKey()
-
-		power := int64(100)
-		if i > 0 {
-			pubKey = ed25519.GenPrivKey().PubKey()
-			power = 1
-		}
-		startTokens := sdk.TokensFromTendermintPower(power)
-
-		msg := staking.NewMsgCreateValidator(
-			sdk.ValAddress(operAddr),
-			pubKey,
-			sdk.NewCoin(sdk.DefaultBondDenom, startTokens),
-			staking.NewDescription(fmt.Sprintf("validator-%d", i+1), "", "", ""),
-			staking.NewCommissionMsg(sdk.ZeroDec(), sdk.ZeroDec(), sdk.ZeroDec()),
-			sdk.OneInt(),
-		)
-		stdSignMsg := txbuilder.StdSignMsg{
-			ChainID: genDoc.ChainID,
-			Msgs:    []sdk.Msg{msg},
-		}
-		sig, err := operPrivKey.Sign(stdSignMsg.Bytes())
-		require.Nil(t, err)
-
-		tx := auth.NewStdTx([]sdk.Msg{msg}, auth.StdFee{}, []auth.StdSignature{{Signature: sig, PubKey: operPrivKey.PubKey()}}, "")
-		genTxs = append(genTxs, tx)
-
-		valConsPubKeys = append(valConsPubKeys, pubKey)
-		valOperAddrs = append(valOperAddrs, sdk.ValAddress(operAddr))
-
-		accAuth := auth.NewBaseAccountWithAddress(sdk.AccAddress(operAddr))
-		accTokens := sdk.TokensFromTendermintPower(150)
-		accAuth.Coins = sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, accTokens)}
-		accs = append(accs, genaccounts.NewGenesisAccount(&accAuth))
-	}
-
-	genesisState := gapp.NewDefaultGenesisState()
-	genDoc.AppState, err = cdc.MarshalJSON(genesisState)
-	require.NoError(t, err)
-	genesisState, err = genutil.SetGenTxsInAppGenesisState(cdc, genesisState, genTxs)
-	require.NoError(t, err)
-
-	// add some tokens to init accounts
-	stakingDataBz := genesisState[staking.ModuleName]
-	var stakingData staking.GenesisState
-	cdc.MustUnmarshalJSON(stakingDataBz, &stakingData)
-	for _, addr := range initAddrs {
-		accAuth := auth.NewBaseAccountWithAddress(addr)
-		accTokens := sdk.TokensFromTendermintPower(100)
-		accAuth.Coins = sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, accTokens)}
-		acc := genaccounts.NewGenesisAccount(&accAuth)
-		accs = append(accs, acc)
-	}
-
-	// distr data
-	distrDataBz := genesisState[distr.ModuleName]
-	var distrData distr.GenesisState
-	cdc.MustUnmarshalJSON(distrDataBz, &distrData)
-	distrData.FeePool.CommunityPool = sdk.DecCoins{sdk.DecCoin{"test", sdk.NewDecFromInt(sdk.NewInt(10))}}
-	distrDataBz = cdc.MustMarshalJSON(distrData)
-	genesisState[distr.ModuleName] = distrDataBz
-
-	// now add the account tokens to the non-bonded pool
-	for _, acc := range accs {
-		accTokens := acc.Coins.AmountOf(sdk.DefaultBondDenom)
-		stakingData.Pool.NotBondedTokens = stakingData.Pool.NotBondedTokens.Add(accTokens)
-	}
-	stakingDataBz = cdc.MustMarshalJSON(stakingData)
-	genesisState[staking.ModuleName] = stakingDataBz
-
-	genaccountsData := genaccounts.NewGenesisState(accs)
-	genaccountsDataBz := cdc.MustMarshalJSON(genaccountsData)
-	genesisState[genaccounts.ModuleName] = genaccountsDataBz
-
-	// mint genesis (none set within genesisState)
-	mintData := mint.DefaultGenesisState()
-	inflationMin := sdk.ZeroDec()
-	if minting {
-		inflationMin = sdk.MustNewDecFromStr("10000.0")
-		mintData.Params.InflationMax = sdk.MustNewDecFromStr("15000.0")
-	} else {
-		mintData.Params.InflationMax = inflationMin
-	}
-	mintData.Minter.Inflation = inflationMin
-	mintData.Params.InflationMin = inflationMin
-	mintDataBz := cdc.MustMarshalJSON(mintData)
-	genesisState[mint.ModuleName] = mintDataBz
-
-	// initialize crisis data
-	crisisDataBz := genesisState[crisis.ModuleName]
-	var crisisData crisis.GenesisState
-	cdc.MustUnmarshalJSON(crisisDataBz, &crisisData)
-	crisisData.ConstantFee = sdk.NewInt64Coin(sdk.DefaultBondDenom, 1000)
-	crisisDataBz = cdc.MustMarshalJSON(crisisData)
-	genesisState[crisis.ModuleName] = crisisDataBz
-
-	// double check inflation is set according to the minting boolean flag
-	if minting {
-		require.Equal(t, sdk.MustNewDecFromStr("15000.0"), mintData.Params.InflationMax)
-		require.Equal(t, sdk.MustNewDecFromStr("10000.0"), mintData.Minter.Inflation)
-		require.Equal(t, sdk.MustNewDecFromStr("10000.0"), mintData.Params.InflationMin)
-	} else {
-		require.Equal(t, sdk.ZeroDec(), mintData.Params.InflationMax)
-		require.Equal(t, sdk.ZeroDec(), mintData.Minter.Inflation)
-		require.Equal(t, sdk.ZeroDec(), mintData.Params.InflationMin)
-	}
-
-	appState, err := codec.MarshalJSONIndent(cdc, genesisState)
-	require.NoError(t, err)
-	genDoc.AppState = appState
-
-	listenAddr, port, err := server.FreeTCPAddr()
-	require.NoError(t, err)
-
-	// NOTE: Need to set this so LCD knows the tendermint node address!
-	viper.Set(client.FlagNode, config.RPC.ListenAddress)
-	viper.Set(client.FlagChainID, genDoc.ChainID)
-	// TODO Set to false once the upstream Tendermint proof verification issue is fixed.
-	viper.Set(client.FlagTrustNode, true)
-
-	node := startTM(t, config, logger, genDoc, privVal, app)
-	require.NoError(t, err)
-
-	tests.WaitForNextHeightTM(tests.ExtractPortFromAddress(config.RPC.ListenAddress))
-	lcd, err := startLCD(logger, listenAddr, cdc, t)
-	require.NoError(t, err)
-
-	tests.WaitForLCDStart(port)
-	tests.WaitForHeight(1, port)
-
-	cleanup = func() {
-		logger.Debug("cleaning up LCD initialization")
-		node.Stop() //nolint:errcheck
-		node.Wait()
-		lcd.Close()
-		os.RemoveAll(config.RootDir)
-	}
-
-	return cleanup, valConsPubKeys, valOperAddrs, port
-}
-
-// startTM creates and starts an in-process Tendermint node with memDB and
-// in-process ABCI application. It returns the new node or any error that
-// occurred.
-//
-// TODO: Clean up the WAL dir or enable it to be not persistent!
-func startTM(
-	t *testing.T, tmcfg *tmcfg.Config, logger log.Logger, genDoc *tmtypes.GenesisDoc,
-	privVal tmtypes.PrivValidator, app abci.Application,
-) *nm.Node {
-
-	genDocProvider := func() (*tmtypes.GenesisDoc, error) { return genDoc, nil }
-	dbProvider := func(*nm.DBContext) (dbm.DB, error) { return dbm.NewMemDB(), nil }
-	nodeKey, err := p2p.LoadOrGenNodeKey(tmcfg.NodeKeyFile())
-	require.NoError(t, err)
-
-	node, err := nm.NewNode(
-		tmcfg,
-		privVal,
-		nodeKey,
-		proxy.NewLocalClientCreator(app),
-		genDocProvider,
-		dbProvider,
-		nm.DefaultMetricsProvider(tmcfg.Instrumentation),
-		logger.With("module", "node"),
-	)
-	require.NoError(t, err)
-
-	err = node.Start()
-	require.NoError(t, err)
-
-	tests.WaitForRPC(tmcfg.RPC.ListenAddress)
-	logger.Info("Tendermint running!")
-
-	return node
-}
-
-// startLCD starts the LCD.
-func startLCD(logger log.Logger, listenAddr string, cdc *codec.Codec, t *testing.T) (net.Listener, error) {
-	rs := lcd.NewRestServer(cdc)
-	registerRoutes(rs)
-	listener, err := tmrpc.Listen(listenAddr, tmrpc.DefaultConfig())
-	if err != nil {
-		return nil, err
-	}
-	go tmrpc.StartHTTPServer(listener, rs.Mux, logger, tmrpc.DefaultConfig()) //nolint:errcheck
-	return listener, nil
-}
-
-// TODO generalize this with the module basic manager
-// NOTE: If making updates here also update cmd/gaia/cmd/gaiacli/main.go
-func registerRoutes(rs *lcd.RestServer) {
-	rpc.RegisterRoutes(rs.CliCtx, rs.Mux)
-	tx.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc)
-	authrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, auth.StoreKey)
-	bankrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	distrrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, distr.StoreKey)
-	stakingrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	slashingrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	govrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, paramsrest.ProposalRESTHandler(rs.CliCtx, rs.Cdc), distrrest.ProposalRESTHandler(rs.CliCtx, rs.Cdc))
-	mintrest.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc)
-}
 
 // Request makes a test LCD test request. It returns a response object and a
 // stringified response body.
@@ -570,19 +159,18 @@ func getTransactionRequest(t *testing.T, port, hash string) (*http.Response, str
 // POST /txs broadcast txs
 
 // GET /txs search transactions
-func getTransactions(t *testing.T, port string, tags ...string) *sdk.SearchTxsResult {
+func getTransactions(t *testing.T, port string, tags ...string) []sdk.TxResponse {
 	var txs []sdk.TxResponse
-	result := sdk.NewSearchTxsResult(0, 0, 1, 30, txs)
 	if len(tags) == 0 {
-		return &result
+		return txs
 	}
 	queryStr := strings.Join(tags, "&")
 	res, body := Request(t, port, "GET", fmt.Sprintf("/txs?%s", queryStr), nil)
 	require.Equal(t, http.StatusOK, res.StatusCode, body)
 
-	err := cdc.UnmarshalJSON([]byte(body), &result)
+	err := cdc.UnmarshalJSON([]byte(body), &txs)
 	require.NoError(t, err)
-	return &result
+	return txs
 }
 
 // ----------------------------------------------------------------------
@@ -954,11 +542,11 @@ func doBeginRedelegation(
 }
 
 // GET /staking/delegators/{delegatorAddr}/delegations Get all delegations from a delegator
-func getDelegatorDelegations(t *testing.T, port string, delegatorAddr sdk.AccAddress) staking.DelegationResponses {
+func getDelegatorDelegations(t *testing.T, port string, delegatorAddr sdk.AccAddress) []staking.Delegation {
 	res, body := Request(t, port, "GET", fmt.Sprintf("/staking/delegators/%s/delegations", delegatorAddr), nil)
 	require.Equal(t, http.StatusOK, res.StatusCode, body)
 
-	var dels staking.DelegationResponses
+	var dels []staking.Delegation
 
 	err := cdc.UnmarshalJSON([]byte(body), &dels)
 	require.Nil(t, err)
@@ -980,7 +568,7 @@ func getDelegatorUnbondingDelegations(t *testing.T, port string, delegatorAddr s
 }
 
 // GET /staking/redelegations?delegator=0xdeadbeef&validator_from=0xdeadbeef&validator_to=0xdeadbeef& Get redelegations filters by params passed in
-func getRedelegations(t *testing.T, port string, delegatorAddr sdk.AccAddress, srcValidatorAddr sdk.ValAddress, dstValidatorAddr sdk.ValAddress) staking.RedelegationResponses {
+func getRedelegations(t *testing.T, port string, delegatorAddr sdk.AccAddress, srcValidatorAddr sdk.ValAddress, dstValidatorAddr sdk.ValAddress) []staking.Redelegation {
 	var res *http.Response
 	var body string
 	endpoint := "/staking/redelegations?"
@@ -993,14 +581,11 @@ func getRedelegations(t *testing.T, port string, delegatorAddr sdk.AccAddress, s
 	if !dstValidatorAddr.Empty() {
 		endpoint += fmt.Sprintf("validator_to=%s&", dstValidatorAddr)
 	}
-
 	res, body = Request(t, port, "GET", endpoint, nil)
 	require.Equal(t, http.StatusOK, res.StatusCode, body)
-
-	var redels staking.RedelegationResponses
+	var redels []staking.Redelegation
 	err := cdc.UnmarshalJSON([]byte(body), &redels)
 	require.Nil(t, err)
-
 	return redels
 }
 
@@ -1050,11 +635,11 @@ func getBondingTxs(t *testing.T, port string, delegatorAddr sdk.AccAddress, quer
 }
 
 // GET /staking/delegators/{delegatorAddr}/delegations/{validatorAddr} Query the current delegation between a delegator and a validator
-func getDelegation(t *testing.T, port string, delegatorAddr sdk.AccAddress, validatorAddr sdk.ValAddress) staking.DelegationResponse {
+func getDelegation(t *testing.T, port string, delegatorAddr sdk.AccAddress, validatorAddr sdk.ValAddress) staking.Delegation {
 	res, body := Request(t, port, "GET", fmt.Sprintf("/staking/delegators/%s/delegations/%s", delegatorAddr, validatorAddr), nil)
 	require.Equal(t, http.StatusOK, res.StatusCode, body)
 
-	var bond staking.DelegationResponse
+	var bond staking.Delegation
 	err := cdc.UnmarshalJSON([]byte(body), &bond)
 	require.Nil(t, err)
 
@@ -1208,8 +793,8 @@ func doSubmitParamChangeProposal(
 		Description: "test",
 		Proposer:    proposerAddr,
 		Deposit:     sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, amount)},
-		Changes: paramscutils.ParamChangesJSON{
-			paramscutils.NewParamChangeJSON("staking", "MaxValidators", "", []byte(`105`)),
+		Changes: []params.ParamChange{
+			params.NewParamChange("staking", "MaxValidators", "", "105"),
 		},
 	}
 
@@ -1217,44 +802,7 @@ func doSubmitParamChangeProposal(
 	require.NoError(t, err)
 
 	resp, body := Request(t, port, "POST", "/gov/proposals/param_change", req)
-	require.Equal(t, http.StatusOK, resp.StatusCode, body)
-
-	resp, body = signAndBroadcastGenTx(t, port, name, pwd, body, acc, client.DefaultGasAdjustment, false)
-	require.Equal(t, http.StatusOK, resp.StatusCode, body)
-
-	var txResp sdk.TxResponse
-	err = cdc.UnmarshalJSON([]byte(body), &txResp)
-	require.NoError(t, err)
-
-	return txResp
-}
-
-func doSubmitCommunityPoolSpendProposal(
-	t *testing.T, port, seed, name, pwd string, proposerAddr sdk.AccAddress,
-	amount sdk.Int, fees sdk.Coins,
-) sdk.TxResponse {
-
-	acc := getAccount(t, port, proposerAddr)
-	accnum := acc.GetAccountNumber()
-	sequence := acc.GetSequence()
-	chainID := viper.GetString(client.FlagChainID)
-	from := acc.GetAddress().String()
-
-	baseReq := rest.NewBaseReq(from, "", chainID, "", "", accnum, sequence, fees, nil, false)
-	pr := distrrest.CommunityPoolSpendProposalReq{
-		BaseReq:     baseReq,
-		Title:       "Test",
-		Description: "test",
-		Proposer:    proposerAddr,
-		Recipient:   proposerAddr,
-		Deposit:     sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, amount)},
-		Amount:      sdk.Coins{sdk.NewCoin("test", sdk.NewInt(5))},
-	}
-
-	req, err := cdc.MarshalJSON(pr)
-	require.NoError(t, err)
-
-	resp, body := Request(t, port, "POST", "/gov/proposals/community_pool_spend", req)
+	fmt.Println(resp)
 	require.Equal(t, http.StatusOK, resp.StatusCode, body)
 
 	resp, body = signAndBroadcastGenTx(t, port, name, pwd, body, acc, client.DefaultGasAdjustment, false)
