+++ lcd_test/lcd_test.go
@@ -1,4 +1,4 @@
-package lcd_test
+package lcdtest
 
 import (
 	"encoding/base64"
@@ -18,6 +18,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/keys"
+	clienttx "github.com/cosmos/cosmos-sdk/client/tx"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/mintkey"
 	"github.com/cosmos/cosmos-sdk/tests"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -52,7 +53,7 @@ func TestVersion(t *testing.T) {
 		t.SkipNow()
 	}
 
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	// node info
@@ -75,14 +76,14 @@ func TestVersion(t *testing.T) {
 }
 
 func TestNodeStatus(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	getNodeInfo(t, port)
 	getSyncStatus(t, port, false)
 }
 
 func TestBlock(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	getBlock(t, port, -1, false)
 	getBlock(t, port, 2, false)
@@ -90,7 +91,7 @@ func TestBlock(t *testing.T) {
 }
 
 func TestValidators(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	resultVals := getValidatorSets(t, port, -1, false)
 	require.Contains(t, resultVals.Validators[0].Address.String(), "cosmosvalcons")
@@ -100,10 +101,11 @@ func TestValidators(t *testing.T) {
 }
 
 func TestCoinSend(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	bz, err := hex.DecodeString("8FA6AB57AD6870F6B5B2E57735F38F2F30E73CB6")
@@ -191,10 +193,11 @@ func TestCoinSend(t *testing.T) {
 }
 
 func TestCoinSendAccAuto(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -216,10 +219,11 @@ func TestCoinSendAccAuto(t *testing.T) {
 }
 
 func TestCoinMultiSendGenerateOnly(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	// generate only
@@ -229,7 +233,7 @@ func TestCoinMultiSendGenerateOnly(t *testing.T) {
 	var stdTx auth.StdTx
 	require.Nil(t, cdc.UnmarshalJSON([]byte(body), &stdTx))
 	require.Equal(t, len(stdTx.Msgs), 1)
-	require.Equal(t, stdTx.GetMsgs()[0].Route(), bank.RouterKey)
+	require.Equal(t, stdTx.GetMsgs()[0].Route(), "bank")
 	require.Equal(t, stdTx.GetMsgs()[0].GetSigners(), []sdk.AccAddress{addr})
 	require.Equal(t, 0, len(stdTx.Signatures))
 	require.Equal(t, memo, stdTx.Memo)
@@ -239,10 +243,11 @@ func TestCoinMultiSendGenerateOnly(t *testing.T) {
 }
 
 func TestCoinSendGenerateSignAndBroadcast(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 
 	defer cleanup()
 	acc := getAccount(t, port, addr)
@@ -265,7 +270,7 @@ func TestCoinSendGenerateSignAndBroadcast(t *testing.T) {
 	var tx auth.StdTx
 	require.Nil(t, cdc.UnmarshalJSON([]byte(body), &tx))
 	require.Equal(t, len(tx.Msgs), 1)
-	require.Equal(t, tx.Msgs[0].Route(), bank.RouterKey)
+	require.Equal(t, tx.Msgs[0].Route(), "bank")
 	require.Equal(t, tx.Msgs[0].GetSigners(), []sdk.AccAddress{addr})
 	require.Equal(t, 0, len(tx.Signatures))
 	require.Equal(t, memo, tx.Memo)
@@ -282,17 +287,19 @@ func TestCoinSendGenerateSignAndBroadcast(t *testing.T) {
 }
 
 func TestEncodeTx(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	res, body, _ := doTransferWithGas(t, port, seed, name1, memo, "", addr, "2", 1, false, false, fees)
 	var tx auth.StdTx
-	require.Nil(t, cdc.UnmarshalJSON([]byte(body), &tx))
+	cdc.UnmarshalJSON([]byte(body), &tx)
 
-	encodedJSON, _ := cdc.MarshalJSON(tx)
+	req := clienttx.EncodeReq{Tx: tx}
+	encodedJSON, _ := cdc.MarshalJSON(req)
 	res, body = Request(t, port, "POST", "/txs/encode", encodedJSON)
 
 	// Make sure it came back ok, and that we can decode it back to the transaction
@@ -315,26 +322,27 @@ func TestEncodeTx(t *testing.T) {
 }
 
 func TestTxs(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	var emptyTxs []sdk.TxResponse
-	txResult := getTransactions(t, port)
-	require.Equal(t, emptyTxs, txResult.Txs)
+	txs := getTransactions(t, port)
+	require.Equal(t, emptyTxs, txs)
 
 	// query empty
-	txResult = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
-	require.Equal(t, emptyTxs, txResult.Txs)
+	txs = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
+	require.Equal(t, emptyTxs, txs)
 
 	// also tests url decoding
-	txResult = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
-	require.Equal(t, emptyTxs, txResult.Txs)
+	txs = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
+	require.Equal(t, emptyTxs, txs)
 
-	txResult = getTransactions(t, port, fmt.Sprintf("action=submit%%20proposal&sender=%s", addr.String()))
-	require.Equal(t, emptyTxs, txResult.Txs)
+	txs = getTransactions(t, port, fmt.Sprintf("action=submit%%20proposal&sender=%s", addr.String()))
+	require.Equal(t, emptyTxs, txs)
 
 	// create tx
 	receiveAddr, resultTx := doTransfer(t, port, seed, name1, memo, pw, addr, fees)
@@ -345,14 +353,14 @@ func TestTxs(t *testing.T) {
 	require.Equal(t, resultTx.TxHash, tx.TxHash)
 
 	// query sender
-	txResult = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	txs = getTransactions(t, port, fmt.Sprintf("sender=%s", addr.String()))
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	// query recipient
-	txResult = getTransactions(t, port, fmt.Sprintf("recipient=%s", receiveAddr.String()))
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	txs = getTransactions(t, port, fmt.Sprintf("recipient=%s", receiveAddr.String()))
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	// query transaction that doesn't exist
 	validTxHash := "9ADBECAAD8DACBEC3F4F535704E7CF715C765BDCEDBEF086AFEAD31BA664FB0B"
@@ -367,10 +375,11 @@ func TestTxs(t *testing.T) {
 }
 
 func TestPoolParamsQuery(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, _ := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, _, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	defaultParams := staking.DefaultParams()
@@ -384,8 +393,8 @@ func TestPoolParamsQuery(t *testing.T) {
 	tokens := sdk.TokensFromTendermintPower(100)
 	freeTokens := sdk.TokensFromTendermintPower(50)
 	initialPool.NotBondedTokens = initialPool.NotBondedTokens.Add(tokens)
-	initialPool.BondedTokens = initialPool.BondedTokens.Add(tokens) // Delegate tx on GaiaAppGenState
-	initialPool.NotBondedTokens = initialPool.NotBondedTokens.Add(freeTokens)
+	initialPool.BondedTokens = initialPool.BondedTokens.Add(tokens)           // Delegate tx on GaiaAppGenState
+	initialPool.NotBondedTokens = initialPool.NotBondedTokens.Add(freeTokens) // freeTokensPerAcc = 50 on GaiaAppGenState
 
 	require.Equal(t, initialPool.BondedTokens, pool.BondedTokens)
 
@@ -395,7 +404,7 @@ func TestPoolParamsQuery(t *testing.T) {
 }
 
 func TestValidatorsQuery(t *testing.T) {
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	require.Equal(t, 1, len(valPubKeys))
@@ -415,7 +424,7 @@ func TestValidatorsQuery(t *testing.T) {
 }
 
 func TestValidatorQuery(t *testing.T) {
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	require.Equal(t, 1, len(valPubKeys))
 	require.Equal(t, 1, len(operAddrs))
@@ -425,11 +434,12 @@ func TestValidatorQuery(t *testing.T) {
 }
 
 func TestBonding(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, _ := CreateAddr(t, name1, pw, kb)
+	addr, _, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
 
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 2, []sdk.AccAddress{addr}, false)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(2, []sdk.AccAddress{addr}, false)
 	tests.WaitForHeight(1, port)
 	defer cleanup()
 
@@ -451,12 +461,12 @@ func TestBonding(t *testing.T) {
 	require.Equal(t, uint32(0), resultTx.Code)
 
 	// query tx
-	txResult := getTransactions(t, port,
+	txs := getTransactions(t, port,
 		fmt.Sprintf("action=delegate&sender=%s", addr),
 		fmt.Sprintf("destination-validator=%s", operAddrs[0]),
 	)
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	// verify balance
 	acc = getAccount(t, port, addr)
@@ -504,12 +514,12 @@ func TestBonding(t *testing.T) {
 	expectedBalance = coins[0]
 
 	// query tx
-	txResult = getTransactions(t, port,
+	txs = getTransactions(t, port,
 		fmt.Sprintf("action=begin_unbonding&sender=%s", addr),
 		fmt.Sprintf("source-validator=%s", operAddrs[0]),
 	)
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	ubd := getUnbondingDelegation(t, port, addr, operAddrs[0])
 	require.Len(t, ubd.Entries, 1)
@@ -541,13 +551,13 @@ func TestBonding(t *testing.T) {
 	)
 
 	// query tx
-	txResult = getTransactions(t, port,
+	txs = getTransactions(t, port,
 		fmt.Sprintf("action=begin_redelegate&sender=%s", addr),
 		fmt.Sprintf("source-validator=%s", operAddrs[0]),
 		fmt.Sprintf("destination-validator=%s", operAddrs[1]),
 	)
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	redelegation := getRedelegations(t, port, addr, operAddrs[0], operAddrs[1])
 	require.Len(t, redelegation, 1)
@@ -575,7 +585,7 @@ func TestBonding(t *testing.T) {
 	// require.Equal(t, sdk.Unbonding, bondedValidators[0].Status)
 
 	// query txs
-	txs := getBondingTxs(t, port, addr, "")
+	txs = getBondingTxs(t, port, addr, "")
 	require.Len(t, txs, 3, "All Txs found")
 
 	txs = getBondingTxs(t, port, addr, "bond")
@@ -589,10 +599,11 @@ func TestBonding(t *testing.T) {
 }
 
 func TestSubmitProposal(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -625,48 +636,12 @@ func TestSubmitProposal(t *testing.T) {
 	require.Equal(t, proposalID, proposer.ProposalID)
 }
 
-func TestSubmitCommunityPoolSpendProposal(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
-	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
-	defer cleanup()
-
-	acc := getAccount(t, port, addr)
-	initialBalance := acc.GetCoins()
-
-	// create proposal tx
-	proposalTokens := sdk.TokensFromTendermintPower(5)
-	resultTx := doSubmitCommunityPoolSpendProposal(t, port, seed, name1, pw, addr, proposalTokens, fees)
-	tests.WaitForHeight(resultTx.Height+1, port)
-
-	// check if tx was committed
-	require.Equal(t, uint32(0), resultTx.Code)
-
-	var proposalID uint64
-	bz, err := hex.DecodeString(resultTx.Data)
-	require.NoError(t, err)
-	cdc.MustUnmarshalBinaryLengthPrefixed(bz, &proposalID)
-
-	// verify balance
-	acc = getAccount(t, port, addr)
-	expectedBalance := initialBalance[0].Sub(fees[0])
-	require.Equal(t, expectedBalance.Amount.Sub(proposalTokens), acc.GetCoins().AmountOf(sdk.DefaultBondDenom))
-
-	// query proposal
-	proposal := getProposal(t, port, proposalID)
-	require.Equal(t, "Test", proposal.GetTitle())
-
-	proposer := getProposer(t, port, proposalID)
-	require.Equal(t, addr.String(), proposer.Proposer)
-	require.Equal(t, proposalID, proposer.ProposalID)
-}
-
 func TestSubmitParamChangeProposal(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -700,10 +675,11 @@ func TestSubmitParamChangeProposal(t *testing.T) {
 }
 
 func TestDeposit(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -744,9 +720,9 @@ func TestDeposit(t *testing.T) {
 	require.Equal(t, expectedBalance.Amount.Sub(depositTokens), acc.GetCoins().AmountOf(sdk.DefaultBondDenom))
 
 	// query tx
-	txResult := getTransactions(t, port, fmt.Sprintf("action=deposit&sender=%s", addr))
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	txs := getTransactions(t, port, fmt.Sprintf("action=deposit&sender=%s", addr))
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	// query proposal
 	totalCoins := sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, sdk.TokensFromTendermintPower(10))}
@@ -759,10 +735,11 @@ func TestDeposit(t *testing.T) {
 }
 
 func TestVote(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -805,9 +782,9 @@ func TestVote(t *testing.T) {
 	expectedBalance = coins[0]
 
 	// query tx
-	txResult := getTransactions(t, port, fmt.Sprintf("action=vote&sender=%s", addr))
-	require.Len(t, txResult.Txs, 1)
-	require.Equal(t, resultTx.Height, txResult.Txs[0].Height)
+	txs := getTransactions(t, port, fmt.Sprintf("action=vote&sender=%s", addr))
+	require.Len(t, txs, 1)
+	require.Equal(t, resultTx.Height, txs[0].Height)
 
 	vote := getVote(t, port, proposalID, addr)
 	require.Equal(t, proposalID, vote.ProposalID)
@@ -846,13 +823,14 @@ func TestVote(t *testing.T) {
 }
 
 func TestUnjail(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, _ := CreateAddr(t, name1, pw, kb)
-	cleanup, valPubKeys, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, _, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, valPubKeys, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
-	// NOTE: any less than this and it fails
+	// XXX: any less than this and it fails
 	tests.WaitForHeight(3, port)
 	pkString, _ := sdk.Bech32ifyConsPub(valPubKeys[0])
 	signingInfo := getSigningInfo(t, port, pkString)
@@ -865,11 +843,11 @@ func TestUnjail(t *testing.T) {
 }
 
 func TestProposalsQuery(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addrs, seeds, names, passwords := CreateAddrs(t, kb, 2)
+	addrs, seeds, names, passwords := CreateAddrs(kb, 2)
 
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addrs[0], addrs[1]}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addrs[0], addrs[1]}, true)
 	defer cleanup()
 
 	depositParam := getDepositParam(t, port)
@@ -995,7 +973,7 @@ func TestProposalsQuery(t *testing.T) {
 }
 
 func TestSlashingGetParams(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/slashing/parameters", nil)
@@ -1007,7 +985,7 @@ func TestSlashingGetParams(t *testing.T) {
 }
 
 func TestDistributionGetParams(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/distribution/parameters", nil)
@@ -1016,10 +994,11 @@ func TestDistributionGetParams(t *testing.T) {
 }
 
 func TestDistributionFlow(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, seed := CreateAddr(t, name1, pw, kb)
-	cleanup, _, valAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, seed, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, valAddrs, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	valAddr := valAddrs[0]
@@ -1092,10 +1071,11 @@ func TestDistributionFlow(t *testing.T) {
 }
 
 func TestMintingQueries(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, _ := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, _, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/minting/parameters", nil)
@@ -1118,10 +1098,11 @@ func TestMintingQueries(t *testing.T) {
 }
 
 func TestAccountBalanceQuery(t *testing.T) {
-	kb, err := keys.NewKeyBaseFromDir(InitClientHome(t, ""))
+	kb, err := keys.NewKeyBaseFromDir(InitClientHome(""))
 	require.NoError(t, err)
-	addr, _ := CreateAddr(t, name1, pw, kb)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	addr, _, err := CreateAddr(name1, pw, kb)
+	require.NoError(t, err)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	bz, err := hex.DecodeString("8FA6AB57AD6870F6B5B2E57735F38F2F30E73CB6")
@@ -1137,5 +1118,4 @@ func TestAccountBalanceQuery(t *testing.T) {
 	res, body = Request(t, port, "GET", fmt.Sprintf("/bank/balances/%s", someFakeAddr), nil)
 	require.Equal(t, http.StatusOK, res.StatusCode, body)
 	require.Contains(t, body, "[]")
-
 }
